+++ src/CharacterInsDelInterface.ts
@@ -1,0 +1,6 @@
+export class CharacterInsDelInterface {
+	constructor(
+		public deletionCost: (c: string) => number,
+		public insertionCost: (c: string) => number,
+	) { }
+}