+++ src/CharacterSubstitutionInterface.ts
@@ -1,0 +1,5 @@
+export class CharacterSubstitutionInterface {
+	constructor(
+		public cost: (c1: string, c2: string) => number,
+	) { }
+}