+++ src/index.ts
@@ -1,1 +1,7 @@
 export * from './JaroWinkler';
+export * from './Levenshtein';
+export * from './CharacterInsDelInterface';
+export * from './CharacterSubstitutionInterface';
+export * from './NormalizedLevenshtein';
+export * from './WeightedLevenshtein';
+export * from './Damerau';