+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 2020-01-04 (0.2.2)
+
+**Feature**
+
+- add damerau-levenstein distance algorithm
+
+# 2020-01-04 (0.2.1)
+
+**Feature**
+
+- add normalized-levenstein distance algorithm
+- add weighted-levenstein distance algorithm
+
 # 2020-01-04 (0.2.0)
 
 **Feature**